+++ src/crazyflie_gazebo/src/math_linear_algebra.cpp
@@ -2,7 +2,7 @@
 
 namespace math {
 
-float clamp(float value, float min, float max) {
+double clamp(double value, double min, double max) {
   if (value < min) return min;
   if (value > max) return max;
   return value;
